[incubator-kie-drools-6848] Restructure drools-test-coverage-jars - Isolate tests which require pre-built jars to test-integration-prebuilt
