--- drools-test-coverage/drools-test-coverage-jars/drools-test-coverage-jars-with-invoker/pom.xml
@@ -1,82 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Licensed to the Apache Software Foundation (ASF) under one
-    or more contributor license agreements.  See the NOTICE file
-    distributed with this work for additional information
-    regarding copyright ownership.  The ASF licenses this file
-    to you under the Apache License, Version 2.0 (the
-    "License"); you may not use this file except in compliance
-    with the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing,
-    software distributed under the License is distributed on an
-    "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-    KIND, either express or implied.  See the License for the
-    specific language governing permissions and limitations
-    under the License.
-
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>drools-test-coverage-jars</artifactId>
-    <groupId>org.drools.testcoverage</groupId>
-    <version>999-SNAPSHOT</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>drools-test-coverage-jars-with-invoker</artifactId>
-
-  <name>Drools :: Test Coverage :: Jars for tests built by invoker</name>
-
-  <properties>
-    <java.module.name>org.drools.compiler.integrationtests.jars.with.invoker</java.module.name>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <org.kie.version>${project.version}</org.kie.version>
-    <path.to.buildlog.directory>${project.parent.parent.basedir}/../${buildlogs.dir}</path.to.buildlog.directory>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.kie</groupId>
-      <artifactId>kie-maven-plugin</artifactId>
-      <exclusions>
-        <exclusion>
-          <groupId>org.apache.maven.resolver</groupId>
-          <artifactId>maven-resolver-util</artifactId>
-        </exclusion>
-      </exclusions>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-invoker-plugin</artifactId>
-        <configuration>
-          <cloneProjectsTo>${project.build.directory}/it</cloneProjectsTo>
-          <logDirectory>${path.to.buildlog.directory}</logDirectory>
-          <streamLogs>true</streamLogs>
-          <goal>clean</goal>
-          <goal>package</goal>
-          <skipInstallation>${skipITs}</skipInstallation>
-          <skipInvocation>${skipITs}</skipInvocation>
-        </configuration>
-        <executions>
-          <execution>
-            <id>integration-test</id>
-            <goals>
-              <goal>run</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-
-    </plugins>
-  </build>
-</project>
