+++ drools-test-coverage/test-compiler-integration/pom.xml
@@ -97,13 +97,6 @@
     </dependency>
 
     <dependency>
-      <groupId>org.drools.testcoverage</groupId>
-      <artifactId>drools-test-coverage-jars</artifactId>
-      <type>pom</type>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
       <groupId>com.thoughtworks.xstream</groupId>
       <artifactId>xstream</artifactId>
       <scope>test</scope>
@@ -193,42 +186,6 @@
       </build>
     </profile>
 
-    <!-- test-compiler-integration requires test jars built by drools-test-coverage-jars -->
-    <!-- This profile makes sure that drools-test-coverage-jars is executed if not-->
-    <!-- See: https://github.com/apache/incubator-kie-drools/pull/6144 -->
-    <profile>
-      <id>missing-test-jar-file</id>
-      <activation>
-        <file>
-          <missing>src/test/resources/surf.jar</missing>
-        </file>
-      </activation>
-      <build>
-        <plugins>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-invoker-plugin</artifactId>
-            <executions>
-              <execution>
-                <id>build-drools-test-coverage-jars</id>
-                <goals>
-                  <goal>run</goal>
-                </goals>
-                <phase>generate-test-resources</phase>
-                <configuration>
-                  <projectsDirectory>${basedir}/../drools-test-coverage-jars</projectsDirectory>
-                  <logDirectory>${path.to.buildlog.directory}</logDirectory>
-                  <goals>
-                    <goal>clean</goal>
-                    <goal>install</goal>
-                  </goals>
-                </configuration>
-              </execution>
-            </executions>
-          </plugin>
-        </plugins>
-      </build>
-    </profile>
   </profiles>
 
 </project>
