+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/DynamicRulesTest.java
@@ -932,122 +932,6 @@ public class DynamicRulesTest {
     @ParameterizedTest(name = "KieBase type={0}")
     @MethodSource("parameters")
     @Timeout(10000)
-    public void testIsolatedClassLoaderWithEnumsPkgBuilder(KieBaseTestConfiguration kieBaseTestConfiguration) throws Exception {
-        try {
-            // Creates first class loader and use it to load fact classes
-            ClassLoader loader1 = new SubvertedClassLoader( new URL[]{getClass().getResource( "/testEnum.jar" )},
-                                                            this.getClass().getClassLoader() );
-            loader1.loadClass( "org.drools.Primitives" );
-            loader1.loadClass( "org.drools.TestEnum" );
-
-            // create a builder with the given classloader
-            Collection<KiePackage> kpkgs = KieBaseUtil.getKieBaseFromClasspathResourcesWithClassLoaderForKieBuilder("test", getClass(), loader1, kieBaseTestConfiguration, "test_EnumSerialization.drl").getKiePackages();
-
-            // adding original packages to a kbase just to make sure they are fine
-            KieBaseConfiguration kbaseConf = RuleBaseFactory.newKnowledgeBaseConfiguration( null, loader1 );
-            final KieModule kieModule = KieUtil.getKieModuleFromResourcesWithClassLoaderForKieBuilder("test", loader1, kieBaseTestConfiguration);
-            InternalKnowledgeBase kbase = (InternalKnowledgeBase) KieBaseUtil.newKieBaseFromReleaseId(kieModule.getReleaseId(), kbaseConf);
-
-            kbase.addPackages( kpkgs );
-            KieSession ksession = kbase.newKieSession();
-            List list = new ArrayList();
-            ksession.setGlobal( "list", list);
-            assertThat(ksession.fireAllRules()).isEqualTo(1);
-            assertThat(list.size()).isEqualTo(1);
-
-            // now, create another classloader and make sure it has access to the classes
-            ClassLoader loader2 = new SubvertedClassLoader( new URL[]{getClass().getResource( "/testEnum.jar" )},
-                                                            this.getClass().getClassLoader() );
-            loader2.loadClass( "org.drools.Primitives" );
-            loader2.loadClass( "org.drools.TestEnum" );
-
-            // create another kbase
-            KieBaseConfiguration kbaseConf2 = RuleBaseFactory.newKnowledgeBaseConfiguration( null,
-                                                                                                  loader2 );
-            final KieModule kieModule2 = KieUtil.getKieModuleFromResourcesWithClassLoaderForKieBuilder("test2", loader2, kieBaseTestConfiguration);
-            InternalKnowledgeBase kbase2 = (InternalKnowledgeBase) KieBaseUtil.newKieBaseFromReleaseId(kieModule2.getReleaseId(), kbaseConf2);
-
-            kbase2.addPackages( kpkgs );
-            ksession = kbase2.newKieSession();
-            list = new ArrayList();
-            ksession.setGlobal( "list", list);
-            assertThat(ksession.fireAllRules()).isEqualTo(1);
-            assertThat(list.size()).isEqualTo(1);
-
-        } catch ( ClassCastException cce ) {
-            cce.printStackTrace();
-            fail( "No ClassCastException should be raised." );
-        }
-    }
-
-    @ParameterizedTest(name = "KieBase type={0}")
-    @MethodSource("parameters")
-    @Timeout(10000)
-    public void testIsolatedClassLoaderWithEnumsContextClassloader(KieBaseTestConfiguration kieBaseTestConfiguration) throws Exception {
-        try {
-            // Creates first class loader and use it to load fact classes
-            ClassLoader loader1 = new SubvertedClassLoader( new URL[]{getClass().getResource( "/testEnum.jar" )},
-                                                            this.getClass().getClassLoader() );
-            loader1.loadClass( "org.drools.Primitives" );
-            loader1.loadClass( "org.drools.TestEnum" );
-
-            // Build it using the current context
-            ClassLoader ccl = Thread.currentThread().getContextClassLoader();
-            Collection<KiePackage> kpkgs;
-            try {
-                Thread.currentThread().setContextClassLoader( loader1 );
-                // create a builder with the given classloader
-                kpkgs = KieBaseUtil.getKieBaseFromClasspathResources("tmp", getClass(), kieBaseTestConfiguration, "test_EnumSerialization.drl").getKiePackages();
-
-                // adding original packages to a kbase just to make sure they are fine
-                InternalKnowledgeBase kbase = (InternalKnowledgeBase) KieBaseUtil.getKieBaseFromKieModuleFromDrl("test", kieBaseTestConfiguration);
-                kbase.addPackages( kpkgs );
-
-                KieSession ksession = kbase.newKieSession();
-                List list = new ArrayList();
-                ksession.setGlobal( "list", list);
-                assertThat(ksession.fireAllRules()).isEqualTo(1);
-                assertThat(list.size()).isEqualTo(1);
-            } finally {
-                Thread.currentThread().setContextClassLoader( ccl );
-            }
-
-            // now, create another classloader and make sure it has access to the classes
-            ClassLoader loader2 = new SubvertedClassLoader( new URL[]{getClass().getResource( "/testEnum.jar" )},
-                                                            this.getClass().getClassLoader() );
-            loader2.loadClass( "org.drools.Primitives" );
-            loader2.loadClass( "org.drools.TestEnum" );
-
-            // set context classloader and use it
-            ccl = Thread.currentThread().getContextClassLoader();
-            try {
-                Thread.currentThread().setContextClassLoader( loader2 );
-                
-                // Note: This test originally serialize/deserialize kpkgs with different context classloaders.
-                // Now exec-model doesn't support package serialization so we may remove this test.
-
-                // create another kbase
-                InternalKnowledgeBase kbase2 = (InternalKnowledgeBase) KieBaseUtil.getKieBaseFromKieModuleFromDrl("test2", kieBaseTestConfiguration);
-                kbase2.addPackages( kpkgs );
-
-                KieSession ksession = kbase2.newKieSession();
-                List list = new ArrayList();
-                ksession.setGlobal( "list", list);
-                assertThat(ksession.fireAllRules()).isEqualTo(1);
-                assertThat(list.size()).isEqualTo(1);
-            } finally {
-                Thread.currentThread().setContextClassLoader( ccl );
-            }
-
-        } catch ( ClassCastException cce ) {
-            cce.printStackTrace();
-            fail( "No ClassCastException should be raised." );
-        }
-    }
-
-    @ParameterizedTest(name = "KieBase type={0}")
-    @MethodSource("parameters")
-    @Timeout(10000)
     public void testDynamicRuleRemovalsSubNetworkAndNot(KieBaseTestConfiguration kieBaseTestConfiguration) throws Exception {
         InternalKnowledgeBase kbase = (InternalKnowledgeBase) KieBaseUtil.getKieBaseFromClasspathResources(getClass(), kieBaseTestConfiguration, "test_DynamicRulesWithNotSubnetwork.drl");
         KieSession ksession = kbase.newKieSession();
