+++ drools-test-coverage/test-compiler-integration/src/test/java/org/drools/mvel/integrationtests/JBRULESTest.java
@@ -18,16 +18,11 @@
  */
 package org.drools.mvel.integrationtests;
 
-import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.jar.JarEntry;
-import java.util.jar.JarInputStream;
 import java.util.stream.Stream;
 
-import org.drools.base.rule.MapBackedClassLoader;
 import org.drools.mvel.compiler.Bar;
 import org.drools.mvel.compiler.Cheese;
 import org.drools.mvel.compiler.FactA;
@@ -47,13 +42,8 @@ import org.kie.api.KieBase;
 import org.kie.api.builder.KieBuilder;
 import org.kie.api.builder.Message.Level;
 import org.kie.api.event.rule.AfterMatchFiredEvent;
-import org.kie.api.io.ResourceType;
 import org.kie.api.runtime.KieSession;
 import org.kie.api.runtime.rule.FactHandle;
-import org.kie.internal.builder.KnowledgeBuilder;
-import org.kie.internal.builder.KnowledgeBuilderConfiguration;
-import org.kie.internal.builder.KnowledgeBuilderFactory;
-import org.kie.internal.io.ResourceFactory;
 import org.mockito.ArgumentCaptor;
 import org.mvel2.MVEL;
 
@@ -315,62 +305,4 @@ public class JBRULESTest {
         ksession.dispose();
     }
 
-    @ParameterizedTest(name = "KieBase type={0}")
-    @MethodSource("parameters")
-    public void testGUVNOR578_2(KieBaseTestConfiguration kieBaseTestConfiguration) throws Exception {
-        // An internal specific test case so not enhanced for executable-model
-        final MapBackedClassLoader loader = new MapBackedClassLoader( this.getClass().getClassLoader() );
-
-        InputStream is = this.getClass().getResourceAsStream("/setter-overload.jar");
-        assertThat(is).as("Make sure to build drools-test-coverage-jars first")
-                .isNotNull();
-        final JarInputStream jis = new JarInputStream(is);
-
-        JarEntry entry;
-        final byte[] buf = new byte[1024];
-        int len;
-        while ( (entry = jis.getNextJarEntry()) != null ) {
-            if ( !entry.isDirectory() ) {
-                final ByteArrayOutputStream out = new ByteArrayOutputStream();
-                while ( (len = jis.read( buf )) >= 0 ) {
-                    out.write( buf,
-                            0,
-                            len );
-                }
-                loader.addResource( entry.getName(),
-                        out.toByteArray() );
-            }
-        }
-
-        final List<JarInputStream> jarInputStreams = new ArrayList<JarInputStream>();
-        jarInputStreams.add(jis);
-
-        final KnowledgeBuilderConfiguration conf = KnowledgeBuilderFactory.newKnowledgeBuilderConfiguration(null, loader);
-        final KnowledgeBuilder kbuilder = KnowledgeBuilderFactory.newKnowledgeBuilder(conf);
-
-        final String header = "import org.drools.compiler.integrationtests.setter.overload.SetterOverload\n";
-
-        kbuilder.add(ResourceFactory.newByteArrayResource(header.getBytes()), ResourceType.DRL);
-        assertThat(kbuilder.hasErrors()).isFalse();
-
-        final String passingRule = "rule \"rule1\"\n"
-                + "dialect \"mvel\"\n"
-                + "when\n"
-                + "SetterOverload( echelle == \"abc\" )"
-                + "then\n"
-                + "end\n";
-
-        final String failingRule = "rule \"rule2\"\n"
-                + "dialect \"mvel\"\n"
-                + "when\n"
-                + "SetterOverload( quotiteRemuneration == 123 , echelle == \"abc\" )"
-                + "then\n"
-                + "end\n";
-
-        kbuilder.add(ResourceFactory.newByteArrayResource(passingRule.getBytes()), ResourceType.DRL);
-        assertThat(kbuilder.hasErrors()).isFalse();
-
-        kbuilder.add(ResourceFactory.newByteArrayResource(failingRule.getBytes()), ResourceType.DRL);
-        assertThat(kbuilder.hasErrors()).isFalse();
-    }
 }
