+++ drools-test-coverage/test-integration-prebuilt/test-integration-prebuilt-jars/kjar-with-invoker/src/it/kie-project-simple/src/main/resources/META-INF/kmodule.xml
