+++ drools-test-coverage/test-integration-prebuilt/test-integration-prebuilt-jars/kjar-with-invoker/src/it/kjar-module-before/src/main/resources/KBase-test/kjar/rule1.drl
