+++ drools-test-coverage/test-integration-prebuilt/test-integration-prebuilt-jars/pom.xml
@@ -19,33 +19,29 @@
     under the License.
 
 -->
+
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <groupId>org.kie</groupId>
-    <artifactId>drools-build-parent</artifactId>
+    <artifactId>test-integration-prebuilt</artifactId>
+    <groupId>org.drools.testcoverage</groupId>
     <version>999-SNAPSHOT</version>
-    <relativePath>../build-parent/pom.xml</relativePath>
   </parent>
+  <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.drools.testcoverage</groupId>
-  <artifactId>drools-test-coverage-parent</artifactId>
+  <artifactId>test-integration-prebuilt-jars</artifactId>
   <packaging>pom</packaging>
+  <name>Drools :: Test Coverage :: Integration tests with prebuilt JARs - JARs</name>
 
-  <name>Drools :: Test Coverage</name>
+  <properties>
+    <parent.project.version>${project.version}</parent.project.version>
+  </properties>
 
   <modules>
-    <module>standalone</module>
-    <module>test-suite</module>
-    <module>test-compiler-integration</module>
-    <module>test-integration-nomvel</module>
-    <module>test-integration-notms</module>
-    <module>test-integration-noxml</module>
-    <module>test-integration-ruleunits</module>
-    <module>test-integration-prebuilt</module>
+    <module>surf</module>
+    <module>testEnum</module>
+    <module>setter-overload</module>
+    <module>kjar-with-invoker</module>
   </modules>
-
 </project>
