+++ drools-test-coverage/test-integration-prebuilt/test-integration-prebuilt-jars/setter-overload/src/main/java/org/drools/compiler/integrationtests/setter/overload/SetterOverload.java
