+++ drools-test-coverage/test-integration-prebuilt/test-integration-prebuilt-jars/surf/src/main/java/org/example/surf/Board.java
