+++ drools-test-coverage/test-integration-prebuilt/test-integration-prebuilt-jars/testEnum/src/main/java/org/drools/Primitives.java
