+++ drools-test-coverage/test-suite/src/test/java/org/drools/testcoverage/functional/KieContainerTest.java
@@ -20,7 +20,6 @@ package org.drools.testcoverage.functional;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Collection;
 
 import org.drools.testcoverage.common.util.TestConstants;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,12 +31,9 @@ import org.kie.api.builder.KieBuilder;
 import org.kie.api.builder.KieFileSystem;
 import org.kie.api.builder.ReleaseId;
 import org.kie.api.builder.model.KieModuleModel;
-import org.kie.api.definition.KiePackage;
-import org.kie.api.definition.rule.Rule;
 import org.kie.api.io.KieResources;
 import org.kie.api.runtime.KieContainer;
 import org.kie.api.runtime.KieSession;
-import org.kie.maven.integration.MavenRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.fail;
@@ -130,52 +126,4 @@ public class KieContainerTest {
         kbuilder.buildAll();
     }
 
-    @Test
-    public void testKieContainerBeforeAndAfterDeployOfSnapshot() throws Exception {
-        // BZ-1007977
-        KieServices ks = KieServices.Factory.get();
-
-        String group = "org.kie.test";
-        String artifact = "test-module";
-        String version = "1.0.0-SNAPSHOT";
-
-        ReleaseId releaseId = ks.newReleaseId(group, artifact, version);
-
-        File kjar = new File("src/test/resources/kjar/kjar-module-before.jar");
-        assertThat(kjar).as("Make sure to build drools-test-coverage-jars first")
-                .exists();
-        File pom = new File("src/test/resources/kjar/pom-kjar.xml");
-        MavenRepository repository = MavenRepository.getMavenRepository();
-        repository.installArtifact(releaseId, kjar, pom);
-
-        KieContainer kContainer = ks.newKieContainer(releaseId);
-        KieBase kbase = kContainer.getKieBase();
-        assertThat(kbase).isNotNull();
-        Collection<KiePackage> packages = kbase.getKiePackages();
-        assertThat(packages).isNotNull();
-        assertThat(packages.size()).isEqualTo(1);
-        Collection<Rule> rules = packages.iterator().next().getRules();
-        assertThat(rules.size()).isEqualTo(2);
-
-        ks.getRepository().removeKieModule(releaseId);
-
-        // deploy new version
-        File kjar1 = new File("src/test/resources/kjar/kjar-module-after.jar");
-        assertThat(kjar1).as("Make sure to build drools-test-coverage-jars first")
-                .exists();
-        File pom1 = new File("src/test/resources/kjar/pom-kjar.xml");
-
-        repository.installArtifact(releaseId, kjar1, pom1);
-
-        KieContainer kContainer2 = ks.newKieContainer(releaseId);
-        KieBase kbase2 = kContainer2.getKieBase();
-        assertThat(kbase2).isNotNull();
-        Collection<KiePackage> packages2 = kbase2.getKiePackages();
-        assertThat(packages2).isNotNull();
-        assertThat(packages2.size()).isEqualTo(1);
-        Collection<Rule> rules2 = packages2.iterator().next().getRules();
-        assertThat(rules2.size()).isEqualTo(4);
-
-        ks.getRepository().removeKieModule(releaseId);
-    }
 }
